refactor: centralize base tsup config
