+++ packages/common/tsup-config.ts
@@ -1,0 +1,62 @@
+import type {Options} from 'tsup'
+import {globSync} from 'glob'
+import {updatePackageExportsPath} from './package-exports-path'
+import {readFileSync} from 'node:fs'
+
+export function getBaseTsupConfig(basePath: string): Options {
+    const entries = globSync(
+        `${basePath}/src/**/*.ts`,
+        {ignore: [`${basePath}/src/**/*.test.ts`]}
+    )
+
+    return {
+        entry: entries,
+        outDir: 'lib',
+        format: ['cjs', 'esm'],
+        bundle: false,
+        dts: true,
+        clean: true,
+        splitting: false,
+        sourcemap: true,
+        outExtension({format}) {
+            return {
+                js: format === 'cjs' ? '.cjs' : '.js',
+            }
+        },
+        async onSuccess() {
+            updatePackageExportsPath(basePath)
+        },
+        define: {
+            __VERSION__: JSON.stringify(JSON.parse(readFileSync('package.json', 'utf8')).version),
+        },
+    }
+}
+
+export function getDevBaseTsupConfig(basePath: string): Options {
+    const entries = globSync(
+        `${basePath}/src/**/*.ts`,
+        {ignore: [`${basePath}/src/**/*.test.ts`]}
+    )
+
+    return {
+        entry: entries,
+        outDir: 'lib',
+        format: 'cjs',
+        bundle: false,
+        clean: true,
+        splitting: false,
+        sourcemap: true,
+        outExtension({format}) {
+            return {
+                js: format === 'cjs' ? '.cjs' : '.js',
+            }
+        },
+        async onSuccess() {
+            updatePackageExportsPath(basePath)
+        },
+        define: {
+            __VERSION__: 'dev',
+        },
+    }
+}
+
