+++ packages/core/tsup.config.dev.ts
@@ -1,26 +1,4 @@
-import {globSync} from 'glob'
 import {defineConfig} from 'tsup'
-import {updatePackageExportsPath} from '../scripts/package-exports-path'
+import {getDevBaseTsupConfig} from '../common/tsup-config'
 
-const entries = globSync('src/**/*.ts', {ignore: ['src/**/*.test.ts']})
-
-export default defineConfig({
-    entry: entries,
-    outDir: 'lib',
-    format: 'cjs',
-    bundle: false,
-    clean: true,
-    splitting: false,
-    sourcemap: true,
-    outExtension({format}) {
-        return {
-            js: format === 'cjs' ? '.cjs' : '.js',
-        }
-    },
-    async onSuccess() {
-        updatePackageExportsPath(__dirname)
-    },
-    define: {
-        __VERSION__: 'dev',
-    },
-})
+export default defineConfig(getDevBaseTsupConfig(__dirname))
