+++ packages/solana/solana-stream/package.json
@@ -26,115 +26,5 @@
     },
     "devDependencies": {
         "typescript": "~5.5.4"
-    },
-    "exports": {
-        "./source": {
-            "import": {
-                "types": "./source.d.ts",
-                "default": "./source.js"
-            },
-            "require": {
-                "types": "./source.d.ts",
-                "default": "./source.cjs"
-            },
-            "types": "./source.d.ts",
-            "default": "./source.js"
-        },
-        "./schema": {
-            "import": {
-                "types": "./schema.d.ts",
-                "default": "./schema.js"
-            },
-            "require": {
-                "types": "./schema.d.ts",
-                "default": "./schema.cjs"
-            },
-            "types": "./schema.d.ts",
-            "default": "./schema.js"
-        },
-        "./query": {
-            "import": {
-                "types": "./query.d.ts",
-                "default": "./query.js"
-            },
-            "require": {
-                "types": "./query.d.ts",
-                "default": "./query.cjs"
-            },
-            "types": "./query.d.ts",
-            "default": "./query.js"
-        },
-        "./instruction": {
-            "import": {
-                "types": "./instruction.d.ts",
-                "default": "./instruction.js"
-            },
-            "require": {
-                "types": "./instruction.d.ts",
-                "default": "./instruction.cjs"
-            },
-            "types": "./instruction.d.ts",
-            "default": "./instruction.js"
-        },
-        ".": {
-            "import": {
-                "types": "./index.d.ts",
-                "default": "./index.js"
-            },
-            "require": {
-                "types": "./index.d.ts",
-                "default": "./index.cjs"
-            },
-            "types": "./index.d.ts",
-            "default": "./index.js"
-        },
-        "./objects/types": {
-            "import": {
-                "types": "./objects/types.d.ts",
-                "default": "./objects/types.js"
-            },
-            "require": {
-                "types": "./objects/types.d.cts",
-                "default": "./objects/types.cjs"
-            },
-            "types": "./objects/types.d.ts",
-            "default": "./objects/types.js"
-        },
-        "./objects/relations": {
-            "import": {
-                "types": "./objects/relations.d.ts",
-                "default": "./objects/relations.js"
-            },
-            "require": {
-                "types": "./objects/relations.d.cts",
-                "default": "./objects/relations.cjs"
-            },
-            "types": "./objects/relations.d.ts",
-            "default": "./objects/relations.js"
-        },
-        "./objects/items": {
-            "import": {
-                "types": "./objects/items.d.ts",
-                "default": "./objects/items.js"
-            },
-            "require": {
-                "types": "./objects/items.d.cts",
-                "default": "./objects/items.cjs"
-            },
-            "types": "./objects/items.d.ts",
-            "default": "./objects/items.js"
-        },
-        "./objects": {
-            "import": {
-                "types": "./objects/index.d.ts",
-                "default": "./objects/index.js"
-            },
-            "require": {
-                "types": "./objects/index.d.cts",
-                "default": "./objects/index.cjs"
-            },
-            "types": "./objects/index.d.ts",
-            "default": "./objects/index.js"
-        }
     }
 }
